Pantalla Itinerarios

+++ Entidades/Itinerario.cs
@@ -9,7 +9,7 @@ namespace Prototipo_CAI;
 public class Itinerario
 {
     public int CodigoItinerario { get; set; }
-    public int CUILCUIT { get; set; }
+    public long CUILCUIT { get; set; }
     public List<Disponibilidad> Disponibilidades { get; set; }
     public bool EstaReservado { get; set; }
     public string NombreCliente { get; set; }
